Refactor(Task): Task가 Primary key로 year + day 조합의 문자열을 사용하도록 수정

+++ app/src/main/java/xyz/ridsoft/harumap/HeatmapAdapter.kt
@@ -43,7 +43,8 @@ class HeatmapAdapter(private val context: Context) :
         DataManager(context)
 
         // Comparator for sorting (by _id)
-        val comparator = Comparator { task1: Task, task2: Task -> task1._id - task2._id }
+        val comparator =
+            Comparator { task1: Task, task2: Task -> task1._id.toInt() - task2._id.toInt() }
         // Get tasks from DataManager and sort
         val tasks = DataManager.tasks.values.sortedWith(comparator)
